+++ examples/tesseract_example.py
