+++ view/search/result_table.py
